+++ csw-location/csw-location-server/src/test/java/csw/location/server/javadsl/JLocationServiceImplTest.java
@@ -9,6 +9,7 @@ import akka.stream.javadsl.Keep;
 import akka.stream.testkit.TestSubscriber;
 import akka.stream.testkit.scaladsl.TestSink;
 import csw.location.api.AkkaRegistrationFactory;
+import csw.location.api.JAkkaRegistrationFactory;
 import csw.location.api.commons.Constants;
 import csw.location.api.javadsl.*;
 import csw.location.api.models.*;
@@ -108,7 +109,7 @@ public class JLocationServiceImplTest extends JUnitSuite {
             InterruptedException {
         int port = 8080;
 
-        AkkaRegistration akkaRegistration = new AkkaRegistrationFactory().make(akkaHcdConnection, actorRef, new Metadata(Map.of("", "")));
+        AkkaRegistration akkaRegistration = JAkkaRegistrationFactory.make(akkaHcdConnection, actorRef, new Metadata(Map.of("", "")));
         HttpRegistration httpRegistration = new HttpRegistration(httpServiceConnection, port, Path);
         TcpRegistration tcpRegistration = new TcpRegistration(tcpServiceConnection, port);
 
@@ -143,7 +144,7 @@ public class JLocationServiceImplTest extends JUnitSuite {
     // DEOPSCSW-39: examples of Location Service
     @Test
     public void testResolveAkkaConnection__DEOPSCSW_13_DEOPSCSW_39() throws ExecutionException, InterruptedException {
-        AkkaRegistration registration = new AkkaRegistrationFactory().make(akkaHcdConnection, actorRef);
+        AkkaRegistration registration = JAkkaRegistrationFactory.make(akkaHcdConnection, actorRef);
         locationService.register(registration).get();
         Assert.assertEquals(registration.location(hostname),
                 locationService.find(akkaHcdConnection).get().orElseThrow());
@@ -176,7 +177,7 @@ public class JLocationServiceImplTest extends JUnitSuite {
     // DEOPSCSW-39: examples of Location Service
     @Test
     public void testAkkaRegistration__DEOPSCSW_13_DEOPSCSW_39() throws ExecutionException, InterruptedException {
-        AkkaRegistration registration = new AkkaRegistrationFactory().make(akkaHcdConnection, actorRef);
+        AkkaRegistration registration = JAkkaRegistrationFactory.make(akkaHcdConnection, actorRef);
 
         locationService.register(registration).get();
         Assert.assertEquals(registration.location(hostname),
@@ -192,7 +193,7 @@ public class JLocationServiceImplTest extends JUnitSuite {
         locationService.register(httpRegistration).get();
 
         //  Register Akka connection
-        AkkaRegistration akkaHcdRegistration = new AkkaRegistrationFactory().make(akkaHcdConnection, actorRef);
+        AkkaRegistration akkaHcdRegistration = JAkkaRegistrationFactory.make(akkaHcdConnection, actorRef);
         locationService.register(akkaHcdRegistration).get();
 
         //  Register Tcp service
@@ -219,14 +220,14 @@ public class JLocationServiceImplTest extends JUnitSuite {
         ComponentId akkaHcdComponentId = new ComponentId(new Prefix(JSubsystem.NFIRAOS, "tromboneHCD"),
                 JComponentType.HCD);
         AkkaConnection akkaHcdConnection = new AkkaConnection(akkaHcdComponentId);
-        AkkaRegistration akkaHcdRegistration = new AkkaRegistrationFactory().make(akkaHcdConnection, actorRef);
+        AkkaRegistration akkaHcdRegistration = JAkkaRegistrationFactory.make(akkaHcdConnection, actorRef);
         locationService.register(akkaHcdRegistration).get();
 
         //  Register Assembly type component
         ComponentId akkaAssemblyComponentId = new ComponentId(new Prefix(JSubsystem.NFIRAOS, "tromboneAssembly"),
                 JComponentType.Assembly);
         AkkaConnection akkaAssemblyConnection = new AkkaConnection(akkaAssemblyComponentId);
-        AkkaRegistration akkaAssemblyRegistration = new AkkaRegistrationFactory().make(akkaAssemblyConnection,
+        AkkaRegistration akkaAssemblyRegistration = JAkkaRegistrationFactory.make(akkaAssemblyConnection,
                 actorRef);
         locationService.register(akkaAssemblyRegistration).get();
 
@@ -234,7 +235,7 @@ public class JLocationServiceImplTest extends JUnitSuite {
         ComponentId akkaContainerComponentId = new ComponentId(new Prefix(JSubsystem.NFIRAOS, "tromboneContainer"),
                 JComponentType.Container);
         AkkaConnection akkaContainerConnection = new AkkaConnection(akkaContainerComponentId);
-        AkkaRegistration akkaContainerRegistration = new AkkaRegistrationFactory().make(akkaContainerConnection,
+        AkkaRegistration akkaContainerRegistration = JAkkaRegistrationFactory.make(akkaContainerConnection,
                 actorRef);
         locationService.register(akkaContainerRegistration).get();
 
@@ -282,7 +283,7 @@ public class JLocationServiceImplTest extends JUnitSuite {
         locationService.register(tcpRegistration).get();
 
         //  Register Akka connection
-        AkkaRegistration akkaRegistration = new AkkaRegistrationFactory().make(akkaHcdConnection, actorRef);
+        AkkaRegistration akkaRegistration = JAkkaRegistrationFactory.make(akkaHcdConnection, actorRef);
         locationService.register(akkaRegistration).get();
 
         Set<Location> locations = Set.of(
@@ -309,7 +310,7 @@ public class JLocationServiceImplTest extends JUnitSuite {
         locationService.register(httpRegistration).get();
 
         // Register Akka connection
-        AkkaRegistration akkaRegistration = new AkkaRegistrationFactory().make(akkaHcdConnection, actorRef);
+        AkkaRegistration akkaRegistration = JAkkaRegistrationFactory.make(akkaHcdConnection, actorRef);
         locationService.register(akkaRegistration).get();
 
         //  filter by Tcp type
@@ -337,9 +338,9 @@ public class JLocationServiceImplTest extends JUnitSuite {
                 ".trombone.hcd3"), JComponentType.HCD));
 
         // Register Akka connection
-        AkkaRegistration akkaRegistration1 = new AkkaRegistrationFactory().make(akkaHcdConnection1, actorRef);
-        AkkaRegistration akkaRegistration2 = new AkkaRegistrationFactory().make(akkaHcdConnection2, actorRef);
-        AkkaRegistration akkaRegistration3 = new AkkaRegistrationFactory().make(akkaHcdConnection3, actorRef);
+        AkkaRegistration akkaRegistration1 = JAkkaRegistrationFactory.make(akkaHcdConnection1, actorRef);
+        AkkaRegistration akkaRegistration2 = JAkkaRegistrationFactory.make(akkaHcdConnection2, actorRef);
+        AkkaRegistration akkaRegistration3 = JAkkaRegistrationFactory.make(akkaHcdConnection3, actorRef);
         locationService.register(akkaRegistration1).get();
         locationService.register(akkaRegistration2).get();
         locationService.register(akkaRegistration3).get();
@@ -440,7 +441,7 @@ public class JLocationServiceImplTest extends JUnitSuite {
                 AkkaTypedExtension.UserActorFactory(typedSystem).spawn(TestActor.behavior(), "test-actor",
                         Props.empty());
 
-        AkkaRegistration registration = new AkkaRegistrationFactory().make(connection, actorRef);
+        AkkaRegistration registration = JAkkaRegistrationFactory.make(connection, actorRef);
         Location location = registration.location(hostname);
         IRegistrationResult registrationResult = locationService.register(registration).get();
         Assert.assertEquals(location, registrationResult.location());